backend/requirements.txt
Django>=5.0,<5.1
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
dj-database-url
google-generativeai
